SQL script to inject inconsistencies in TPC-H data
This script takes the .tbl files generated using the DBGen tool of TPC-H and inserts data in an SQL Server database. It injects inconsistencies with respect to primary keys based on the parameters P (inconsistency percentage) and x (number of intended key-equal groups of size > 1). It also creates supplementary tables that are used by CAvSAT/AggCAvSAT.